+++ Source/AccelByteUe4Sdk/Private/Core/AccelByteCredentials.cpp
@@ -63,8 +63,7 @@ void Credentials::ForgetAll()
 
 void Credentials::SetClientCredentials(const ESettingsEnvironment Environment)
 {
-	//Only read ClietId & ClientSecret at the end of scope
-	FReadScopeLock ReadLock(CredentialAccessLock);
+	//Synchronization is handled by BaseCredentials::SetClientCredentials() member method
 
 	FString SectionPath;
 	switch (Environment)
@@ -92,16 +91,19 @@ void Credentials::SetClientCredentials(const ESettingsEnvironment Environment)
 		SectionPath = TEXT("/Script/AccelByteUe4Sdk.AccelByteSettings");
 		break;
 	}
-
-	if (GConfig->GetString(*SectionPath, TEXT("ClientId"), ClientId, GEngineIni))
+	
+	FString ClientIdFromConfig;
+	FString ClientSecretFromCConfig;
+	if (GConfig->GetString(*SectionPath, TEXT("ClientId"), ClientIdFromConfig, GEngineIni))
 	{
-		GConfig->GetString(*SectionPath, TEXT("ClientSecret"), ClientSecret, GEngineIni);
+		GConfig->GetString(*SectionPath, TEXT("ClientSecret"), ClientSecretFromCConfig, GEngineIni);
 	}
 	else
 	{
-		GConfig->GetString(DefaultSection, TEXT("ClientId"), ClientId, GEngineIni);
-		GConfig->GetString(DefaultSection, TEXT("ClientSecret"), ClientSecret, GEngineIni);
+		GConfig->GetString(DefaultSection, TEXT("ClientId"), ClientIdFromConfig, GEngineIni);
+		GConfig->GetString(DefaultSection, TEXT("ClientSecret"), ClientSecretFromCConfig, GEngineIni);
 	}
+	SetClientCredentials(ClientIdFromConfig, ClientSecretFromCConfig);
 }
 
 void Credentials::SetAuthToken(const FOauth2Token& NewAuthToken, float CurrentTime)